+++ src/test/java/org/opensearch/knn/index/util/NativeLibraryTests.java
@@ -64,5 +64,15 @@ public class NativeLibraryTests extends KNNTestCase {
         ) {
             super(methods, scoreTranslation, currentVersion, extension);
         }
+
+        @Override
+        public Float distanceToRadialThreshold(Float distance, SpaceType spaceType) {
+            return 0.0f;
+        }
+
+        @Override
+        public Float scoreToRadialThreshold(Float score, SpaceType spaceType) {
+            return 0.0f;
+        }
     }
 }